move not specific functions from utils
