+++ src/utils.py
@@ -1,6 +1,5 @@
 import cv2
 import numpy as np
-import segmentation_models_pytorch as smp
 import torch
 
 
@@ -8,30 +7,6 @@ def get_device(device: str):
     if device is None:
         return torch.device('cuda') if torch.cuda.is_available() else torch.device('cpu')
     return torch.device(device)
-
-
-def get_mask_rcnn(weights, box_conf_th: float, nms_th: float, device):
-    model = torch.load(weights, map_location=device)
-    model.roi_heads.score_thresh = box_conf_th
-    model.roi_heads.nms_thresh = nms_th
-    model.eval()
-    return model
-
-
-def get_unet(weights, device):
-
-    CLASSES = ['coal', 'bound', 'background']
-    model = smp.Unet(
-        encoder_name='efficientnet-b0',
-        encoder_weights='imagenet',
-        classes=len(CLASSES),
-        activation='softmax',
-    )
-    model_state_dict = torch.load(weights, map_location=device)
-    model.load_state_dict(model_state_dict)
-    model = model.to(device)
-    model.eval()
-    return model
 
 
 def get_contours(mask):
